+++ R/exit-handlers.R
@@ -1,8 +1,22 @@
 
 `_renv_exit_handlers` <- new.env(parent = emptyenv())
+`_renv_exit_handlers_id` <- 1L
+
+renv_exit_handlers_key <- function(envir) {
+
+  # check for existing id
+  id <- attr(envir, "__renv_exit_handlers_id__", exact = TRUE)
+  if (!is.null(id))
+    return(id)
+
+  # set id if null
+  `_renv_exit_handlers_id` <<- `_renv_exit_handlers_id` + 1L
+  attr(envir, "__renv_exit_handlers_id__") <- as.character(`_renv_exit_handlers_id`)
+
+}
 
 renv_exit_handlers_get <- function(envir) {
-  key <- format(envir)
+  key <- renv_exit_handlers_key(envir)
   `_renv_exit_handlers`[[key]]
 }
 
@@ -14,13 +28,13 @@ renv_exit_handlers_set <- function(envir, handlers) {
     do.call(base::on.exit, list(substitute(call), TRUE), envir = envir)
   }
 
-  key <- format(envir)
+  key <- renv_exit_handlers_key(envir)
   `_renv_exit_handlers`[[key]] <- handlers
 
 }
 
 renv_exit_handlers_remove <- function(envir) {
-  key <- format(envir)
+  key <- renv_exit_handlers_key(envir)
   `_renv_exit_handlers`[[key]] <- NULL
 }
 
@@ -35,4 +49,8 @@ renv_exit_handlers_add <- function(envir, handler) {
   handlers <- c(list(handler), renv_exit_handlers_get(envir))
   renv_exit_handlers_set(envir, handlers)
   handler
+}
+
+renv_exit_handlers_task <- function() {
+  `_renv_exit_handlers_id` <<- 1L
 }